test(api): refactor to extract login logic into 'loginUser' util function -- use in all methods except for the primary login test
